Turn qGetLowPointerBits() and qClearLowPointerBits() into a template
The previous implementation had a bug that was triggering the Q_ASSERT
at runtime. Unfortunately that was happening while releasing the mutex
that is needed in order to print the Q_ASSERT error message so the
program deadlocked. By making this a template the mask can be a
compile-time constant which means I can add a static_assert that the mask
value is sensible (i.e. don't use more than the low 5 pointer bits).