Make QArrayData work in the purecap ABI
This is slightly tricky since we can no longer store an offset to the data
in all cases (since that would require deriving from an abient capability).
Instead we can store a __uintcap_t that is either a valid pointer for the
`fromRawData()` case or an offset if this is a contiguous allocation.
Since QArrayData uses a single malloc() for the header and the string data
we can just have an offset into the data to get at the string and then use
a CSetBounds to limit it to just the string. The trickier case is the
global arrays created by QStringLiteral() or QByteArrayLiteral().
Luckliy, it turns out that the compiler sets the object size to the full
`QStaticArrayData` when creating global array literal objects so the offset
variant also works for those globals.