Add an extra character to QArrayData::data() bounds for global constants
QByteArrayLiterals and the tables generated by moc all contain
null-terminated strings. Since they are global static data the alloc
member is 0 and size == sizeof(str) - 1. In order to use these for printf
we need to include the null terminator in the result of data().
This also fixes the sharedNull() in a more generic way.