Allow additional QStandardPaths data directories on macOS
On macOS QStandardPaths::GenericDataLocation only contains
"~/Library/Application Support" and "/Library/Application Support".
Package managers such as homebrew or KDE craft install into a custom
prefix and do not/cannot write to either of these directories. This makes
it impossible for them to share the generic data between different
applications. The current workaround used e.g. by homebrew-kde is to create
symlinks for every file/directory that is needed in /usr/local/share/ in
the user's "~/Library/Application Support" directory. However, this causes
issues if you forget to create the appropriate link and makes it difficult
for the user to override a data file from the system-wide installation with
a file in the home directory.

This commit introduces a ./configure option `-additional-datadir`.
This option can be passed multiple times to create a list of additional
directories that QStandardPaths searches before looking in
"/Library/Application Support".

[ChangeLog][QtCore][QStandardPaths] Added configure option
-additional-datadir to specify additional data directories.

Change-Id: I23b046fcc9520181286fe300827645cd608b0201