[UPSTREAM?] QFileSystemMetaData: Avoid copying arbitrary stack data
This class is copied using a compiler-generated copy constructor (i.e.
a call to memcpy) when creating a new QFileInfo object. This can copy
uninitialized data from the stack to the new object. While this is
generally harmless since it will be overwritten anyway shortly afterwards,
it did trigger the misaligned capability memcpy detection in the CheriBSD
MIPS libc (copying a capability from the stack to an 8-byte aligned
region). Initialize all members to silence this diagnostic and avoid
copies of uninitialized memory.

Change-Id: I009048e2d2cf5c327fc2afdb40a5516b37077a90