Enable QByteArray::number((u)intcap_t)
Now that we no longer support the offset-based interpretation of
uintcap_t there is no need to delete these overloads.
When I did the initial port, offset interpretatation was still the
default so using the uintcap overloads would print the difference
between the capability base and offset (zero in many cases) instead
of the address. This will allow removing lots of local diffs that add
casts in QByteArray::number() calls.

Change-Id: I514bb0dc3f986194500993be4200a4691a700264