Add (u)intcap versions of QString::number and QString::arg
These are needed by QtDeclarative. Now that we no longer support the
offset-based interpretation of uintcap_t there is no need to delete
these overloads. When I did the initial port, offset interpretatation
was still the default so using the uintcap overloads would print the
difference between the capability base and offset (zero in many cases)
instead of the address. This will allow removing lots of local diffs
that add casts in QString::number() calls.

Change-Id: I7ad3ccb4ae2f7d29719e745e8604ca2d68b7a8d2
Co-authored-by: pentelbart <46610525+pentelbart@users.noreply.github.com>