Avoid an out-of-bounds read in rawTypeNameFromTypeInfo()
The NUL terminator is was not part of the QByteArray::fromRawData(), so
calling strlen() on the raw char* later resulted in a bounds violation.
We could work around this by creating a copy of the string data, but in
this patch I chose to add QByteArray::fromNulTerminatedRawData() instead.
This behaves like QByteArray::fromRawData(), but also includes the trailing
'\0' in the capability bounds, but does not include it in `size`.
Passing a NULL pointer to this function is invalid as is passing data
where ptr[size] != '\0'.

Change-Id: Id259e8f94426f8c6b720ff9d8bde14c5559658bc