Ensure that QString().data() has an in-bounds nul character
There are a few cases where an empty QString is passed to a function that
expects a zero-terminated ushort*. The case where I found this was a
qWarning("... %ls\n", qUtf16Printable(this->foo)) call with an empty
string argument. This then crashes inside QString::vasprintf() if we
don't include the nul char in the capability bounds.

Change-Id: I589a167e9b46d4a83fa9a87b68bda32a9568325a