Fix QDBusInterface crashing with empty path on construction
On some machines having an empty path passed to QDBusInterface
would cause a crash.

This happened because the code created a QDBusMessage and manually
marked it as validated when it was not. The validation would not
pass for this object.

Change-Id: I496dd922fa64353399655a1e84996b99990f5879
Pick-to: 6.4 6.3 6.2 5.15
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>
(cherry picked from commit 4d60ba61dca7d50c8eaf9f4525cb9565b363ca81)

* asturmlechner 2022-07-02: Replace _L1 with QLatin1String() for 5.15