Import the 5.15.8p157 patch collection
Upstream overwrites the kde/5.15 branch history when rebasing it onto
a newer Qt version and it's not possible to reproduce the branch version
that existed at the time a specific patch collection release was
published.

The import is against the v5.15.8-lts-lgpl branch using a distfile
archive cached by the FreeBSD ports collection.